+++ ElvUI_MerathilisUI/Core/Changelog/7.09.lua
@@ -1,8 +1,12 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[709] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "01.02.2026",
+	FIXES = {
+		"[Tooltip]: Fixed an nil error for achievements",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[Profile]: Added random hearthstone to the WindTools game bar setting",
+	},
 }