+++ ElvUI_MerathilisUI/Modules/Profiles/ElvUI/WindTools.lua
@@ -55,6 +55,7 @@ function module:LoadWindToolsProfile()
 	db["misc"]["automation"]["hideBagAfterEnteringCombat"] = true
 	db["misc"]["automation"]["hideWorldMapAfterEnteringCombat"] = true
 	db["misc"]["gameBar"]["enable"] = true
+	db["misc"]["gameBar"]["hearthstone"]["left"] = "RANDOM"
 	db["misc"]["spellActivationAlert"]["enable"] = true
 	db["quest"]["turnIn"]["enable"] = false
 	db["social"]["chatText"]["removeBrackets"] = true