+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Changes ###
 
-  * [Install]: Some install and update fixes
+  * [Tooltip]: Fixed an nil error for achievements
+  * [Profile]: Added random hearthstone to the WindTools game bar setting